Add files via upload
I just worked on a dataset of 1000 unicorn companies with a valuation of a minimum of $1 billion each. I created a table from the dataset in Excel. I cleaned the dataset and prepared it for visualization in Excel. I created some Pivot tables from which I created Pivot Charts that I eventually displayed in a Dashboard that I created from scratch in Microsoft Excel. The unique thing about this project is that when I have a project to work on and by providence, visualization platforms such as Power BI and Tableau are not available, a professional Dashboard can still be produced from which insight would be generated for informed decision-making. It is awesome!